Fixed Chrome bug with file loading
Fixed a bug where Google Chrome would not allow you to load the same file twice in a row and did not open a new tab for the second attempt, even if the first tab was closed. Every file (even if it is the same file) is now loaded into a new tab in the editor.